+++ dss/api/bundles/__init__.py
@@ -15,10 +15,10 @@ from dss import DSSException, dss_handler, DSSForbiddenException
 from dss.api.search import PerPageBounds
 from dss.config import Config, Replica
 from dss.storage.blobstore import idempotent_save, test_object_exists, ObjectTest
-from dss.storage.bundles import get_bundle_manifest, save_bundle_manifest, enumerate_avaliable_bundles
+from dss.storage.bundles import get_bundle_manifest, save_bundle_manifest
 from dss.storage.checkout import CheckoutError, TokenError
 from dss.storage.checkout.bundle import get_dst_bundle_prefix, verify_checkout
-from dss.storage.identifiers import BundleTombstoneID, FileFQID, BUNDLE_PREFIX
+from dss.storage.identifiers import BundleTombstoneID, TOMBSTONE_SUFFIX, BUNDLE_PREFIX
 from dss.storage.hcablobstore import BundleFileMetadata, BundleMetadata, FileMetadata
 from dss.util import UrlBuilder, security, hashabledict
 from dss.util.version import datetime_to_version_format
@@ -151,6 +151,48 @@ def enumerate(replica: str,
     :param search_after: used to page searches, should not be set by the user.
     """
 
+    def enumerate_avaliable_bundles(replica: str = None,
+                                    prefix: typing.Optional[str] = None,
+                                    per_page: int = PerPageBounds.per_page_max,
+                                    search_after: typing.Optional[str] = None,
+                                    token: typing.Optional[str] = None):
+        """
+        :returns: returns dictionary with bundles that are available, provides context of cloud providers internal pagination
+                 mechanism.
+        :rtype: dictionary
+        """
+        kwargs = dict(bucket=Replica[replica].bucket, prefix=prefix, k_page_max=per_page)
+        if search_after:
+            kwargs['start_after_key'] = search_after
+        if token:
+            kwargs['token'] = token
+
+        storage_handler = Config.get_blobstore_handle(Replica[replica])
+        prefix_iterator = storage_handler.list_v2(**kwargs)  # note dont wrap this in enumerate() it looses the token
+        keys = dict()  # type: dict
+        uuid_list = list()
+        total_keys = 0
+
+        for key, meta in prefix_iterator:
+            uuid, version = key.split('.', 1)
+            uuid = uuid.split(f'{BUNDLE_PREFIX}/')[1]
+            search_after = key
+            if not version.endswith(TOMBSTONE_SUFFIX):
+                keys.setdefault(uuid, []).append(version)
+                total_keys += 1
+            elif TOMBSTONE_SUFFIX == version:
+                if uuid in keys:
+                    total_keys -= len(keys[uuid])
+                    del keys[uuid]
+            if total_keys >= per_page:
+                break
+
+        for uuid, versions in keys.items():
+            for version in versions:
+                uuid_list.append(dict(uuid=uuid, version=version))
+        token = getattr(prefix_iterator, 'token', None)
+        return dict(search_after=search_after, bundles=uuid_list, token=token, page_count=len(uuid_list))
+
     if prefix:
         search_prefix = f'{BUNDLE_PREFIX}/{prefix}'
     else:
@@ -172,6 +214,7 @@ def enumerate(replica: str,
         del payload['token']
         del payload['search_after']
         response = make_response(jsonify(payload), requests.codes.ok)
+        response.headers['X-OpenAPI-Pagination'] = 'false'
     else:
         next_url = UrlBuilder(request.url)
         next_url.replace_query("search_after", payload['search_after'])
@@ -180,6 +223,8 @@ def enumerate(replica: str,
         payload.update(dict(has_more=True, token=payload['token'], link=f'{next_url}'))
         response = make_response(jsonify(payload), requests.codes.partial)
         response.headers['Link'] = link
+        response.headers['X-OpenAPI-Pagination'] = 'true'
+        response.headers['X-OpenAPI-Paginated-Content-Key'] = 'bundles'
     return response
